Include necessary dependency.
I should have declared this dependency a while ago. I have been using it already for a while. This fixes the error "could not load module". I didn't get this error before because of the way my machine was set up, but I noticed it when using haskell-stack on a different machine.